requirements.txt
torch
torchvision
tabulate
xlsxwriter
pandas
numpy
pillow
kornia
scikit-learn